+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Field `address_cache` added to the `socket_server` input.
+- Field `read_header` added to the `amqp_1` input.
 - All inputs with a `codec` field now support a new field `scanner` to replace it. Scanners are more powerful as they are configured in a structured way similar to other component types rather than via a single string field, for more information [check out the scanners page](https://www.benthos.dev/docs/components/scanners/about).
 - New `diff` and `patch` Bloblang methods.
 
@@ -15,6 +16,10 @@ All notable changes to this project will be documented in this file.
 
 - The `javascript` processor now handles module imports correctly.
 - Bloblang `if` statements now provide explicit errors when query expressions resolve to non-boolean values.
+
+### Fixed
+
+- Metadata fields of `amqp_1` input were always empty due to type mismatch.
 
 ### Changed